move Dialog into components directory
