+++ src/components/nabla/Pools/Swap/SwapPoolModals.tsx
@@ -1,6 +1,6 @@
 import { FunctionalComponent } from 'preact';
 import { ModalTypes, useModal } from '../../../../services/modal';
-import { Dialog } from '../../../../pages/staking/dialogs/Dialog';
+import { Dialog } from '../../../Dialog';
 import { SwapPoolColumn } from './columns';
 import AddLiquidity from './AddLiquidity';
 import Redeem from './Redeem';
