+++ src/pages/gas/GasSuccessDialog.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'preact/compat';
 import { Button } from 'react-daisyui';
 import SuccessDialogIcon from '../../assets/dialog-status-success';
-import { Dialog } from '../staking/dialogs/Dialog';
+import { Dialog } from '../../components/Dialog';
 
 interface DialogProps {
   visible: boolean;
