+++ src/pages/spacewalk/transactions/TransferDialog/TransferDialog.tsx
@@ -4,14 +4,14 @@ import { JSXInternal } from 'preact/src/jsx';
 import { Divider } from 'react-daisyui';
 
 import { useGlobalState } from '../../../../GlobalStateProvider';
-import { CopyableAddress } from '../../../../components/PublicKey';
 import { deriveShortenedRequestId } from '../../../../helpers/spacewalk';
 import { convertRawHexKeyToPublicKey } from '../../../../helpers/stellar';
 import { toTitle } from '../../../../helpers/string';
 import { useVaultRegistryPallet } from '../../../../hooks/spacewalk/useVaultRegistryPallet';
 import { nativeToDecimal } from '../../../../shared/parseNumbers/metric';
+import { CopyableAddress } from '../../../../components/PublicKey';
+import { Dialog } from '../../../../components/Dialog';
 import { TTransfer } from '../TransactionsColumns';
-import { Dialog } from '../../../staking/dialogs/Dialog';
 
 export interface BaseTransferDialogProps {
   id: string;
